@@ -1,30 +1,27 @@
 name: yolov5ncnn_example
 publish_to: none
 description: Demonstrates how to use the yolov5ncnn plugin.
-environment: 
-  sdk: '>=2.19.1 <3.0.0'
+environment:
+  sdk: ">=3.1.3 <4.0.0"
 
-dependencies: 
+dependencies:
   cupertino_icons: ^1.0.2
-  bitmap: ^0.1.3
+  bitmap: ^0.2.0
   image_picker: 0.8.7
   camera: 0.10.3+2
   path_provider: 2.0.14
   image: 3.3.0
-  flutter: 
+  flutter:
     sdk: flutter
-  yolov5ncnn: 
+  yolov5ncnn:
     path: ../
   flutter_isolate: ^2.0.4
   isolate_flutter: ^3.0.0
 
-dev_dependencies: 
-  flutter_lints: ^2.0.0
-  flutter_test: 
-    sdk: flutter
+dev_dependencies:
+  flutter_lints: ^3.0.0
 
-flutter: 
-  assets: 
+flutter:
+  assets:
     - images/
   uses-material-design: true
-
