@@ -1,3 +1,9 @@
+/*
+ * @Date: 2023-03-15 14:49:49
+ * @LastEditors: dengxin 994386508@qq.com
+ * @LastEditTime: 2024-01-15 11:17:40
+ * @FilePath: /yolov5ncnn/lib/yolov5ncnn.dart
+ */
 import 'dart:convert';
 
 import 'package:bitmap/bitmap.dart';
@@ -11,12 +17,12 @@ class Yolov5ncnn {
     return await methodChannel.invokeMethod<bool>('init');
   }
 
-  Future<List<YolovResult>> detect(Bitmap bitmap, bool use_gpu) async {
+  Future<List<YolovResult>> detect(Bitmap bitmap, bool useGpu) async {
     List<YolovResult> list = [];
     try {
       var res = await methodChannel.invokeMethod('detect', {
         "bitmap": bitmap.content,
-        "use_gpu": use_gpu,
+        "use_gpu": useGpu,
         "width": bitmap.width,
         "height": bitmap.height
       });
